+++ client/package.json
@@ -1,6 +1,7 @@
 {
   "name": "@harmonia/client",
   "type": "module",
+  "root": true,
   "scripts": {
     "dev": "vite --host",
     "build": "vite build",
@@ -20,7 +21,6 @@
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-jsx-a11y": "^6.8.0",
-    "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
     "prettier": "^3.2.5",
