+++ client/src/assets/images/Images.js
@@ -1,2 +1,1 @@
-export {default as Bannière} from "./Bannière.svg";
-export {default as CarouselOne} from "./CarouselOne.jpg";
+export {default as banniere} from "./logos/banniere.svg";
