+++ client/src/components/Header/Header.css
@@ -5,11 +5,11 @@ header .logo {
 .burgermenu {
   color: white;
 }
-a.burgermenu_link {
+a.burgermenu-link {
   font-size: 1rem;
   padding-left: 0;
 }
-.burgermenu_links {
+.burgermenu-links {
   display: flex;
   flex-direction: column;
   position: fixed;
@@ -21,55 +21,27 @@ a.burgermenu_link {
   visibility: hidden;
   background-color: rgba(123, 123, 123, 0.9);
   left: 0;
-  transition: all 0.1s ease-out;
+  transition: all 0.1s ease-in-out;
 }
-.show-burger .burger-bar,
-.show-burger .burger-bar::before,
-.show-burger .burger-bar::after {
-  background-color: white;
-}
-.show-burger .burgermenu_links {
+.show-burger .burgermenu-links {
   width: 17rem;
   visibility: visible;
 }
-.burgermenu_link {
+.burgermenu-link {
   display: block;
   padding: 1.5rem;
   color: inherit;
 }
 .burger-bar {
-  width: 40px;
-  height: 40px;
-  background-color: transparent;
-  border: none;
-  color: inherit;
-  display: block;
-  position: fixed;
-  top: 1rem;
-  left: 1rem;
-}
-.burger-bar:hover {
-  cursor: pointer;
-}
-.burger-bar,
-.burger-bar::before,
-.burger-bar::after {
-  display: block;
-  width: 40px;
-  height: 4px;
-  position: relative;
-  border-radius: 3px;
-  background-color: #7b7b7b;
-}
-.burger-bar::before,
-.burger-bar::after {
-  content: "";
+  z-index: 1;
   position: absolute;
-  left: 0;
+  border: none;
+  padding: 1%;
+  background-color: transparent;
 }
-.burger-bar::before {
-  transform: translateY(-12px);
+.hide-burger .burger-bar path {
+  fill: var(--secondary-color);
 }
-.burger-bar::after {
-  transform: translateY(12px);
+.show-burger .burger-bar path {
+  fill: white;
 }
