--- client/src/components/HeroCard/HeroCard.css
@@ -1,16 +1,0 @@
-.hero-card {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  width: 23%;
-  margin: 1%;
-}
-.hero-card img {
-  width: 100%;
-  height: 200px;
-  object-fit: cover;
-  border-radius: 25px;
-}
-.card-description {
-  font-size: small;
-}
